requirements.txt
# --- API ---
fastapi
uvicorn

# --- Database ---
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9

# --- Data ---
pandas>=2.0.0
numpy>=1.24.0
openpyxl>=3.1.0

# --- Matching / NLP ---
rapidfuzz>=3.0.0
scikit-learn>=1.3.0
spacy>=3.6.0
sentence-transformers>=2.2.0

# --- LLM ---
openai>=1.0.0
# see note below for Gemini

# --- Config ---
python-dotenv>=1.0.0
pydantic>=2.0.0
google-generativeai
